+++ .changeset/tempo-abi-groups.md
@@ -1,0 +1,5 @@
+---
+'viem': patch
+---
+
+Added `Abis.core`, `Abis.earn`, and `Abis.zone` contract groups.
