+++ .changeset/v3-tempo-abis-all.md
@@ -1,0 +1,10 @@
+---
+'viem': minor
+---
+
+**Breaking (`viem/tempo`):** Renamed `Abis.abis` to `Abis.all` and expanded the aggregate with Earn and Zone ABIs.
+
+```diff
+-const abis = Abis.abis
++const abis = Abis.all
+```
