+++ .env.example
@@ -11,4 +11,5 @@ VITE_TEMPO_ENV=localnet                     # node environment. values: localnet
 VITE_TEMPO_LOG=false                        # whether to enable logs, if VITE_TEMPO_ENV=localnet. values: true|debug|error|trace|warn|info
 VITE_TEMPO_TAG=latest                       # docker image tag, if VITE_TEMPO_ENV=localnet.
 VITE_TEMPO_ZONES=false                      # whether to run local Zone tests. values: true|false
+VITE_TEMPO_ZONE_RPC_URL=                    # remote Zone RPC URL
 VITE_TEMPO_HTTP_LOG=false                   # whether to enable logs on HTTP RPC requests. values: true|false
