+++ site/snippets/tempo/zones.config.ts
@@ -1,11 +1,10 @@
 // [!region setup]
 import { privateKeyToAccount } from 'viem/accounts'
-import { createClient } from 'viem/tempo'
-import { http, zoneModerato } from 'viem/tempo/zones'
+import { createClient, http, Zone } from 'viem/tempo'
 
 export const client = createClient({
   account: privateKeyToAccount('0x...'),
-  chain: zoneModerato(7),
-  transport: http(),
+  chain: Zone.a,
+  transport: http('https://rpc.example.com'),
 })
 // [!endregion setup]
