+++ src/tempo/actions/zone.test-d.ts
@@ -7,7 +7,7 @@ import { createClient } from '../../clients/createClient.js'
 import { custom } from '../../clients/transports/custom.js'
 import { decorator } from '../Decorator.js'
 import type { TransactionReceipt } from '../Transaction.js'
-import { zoneModerato } from '../zones/index.js'
+import * as Zone from '../Zone.js'
 import * as zoneActions from './zone.js'
 
 const transport = custom({
@@ -26,7 +26,7 @@ const publicClient = createClient({
 })
 const zoneClient = createClient({
   account: '0x0000000000000000000000000000000000000001',
-  chain: zoneModerato(7),
+  chain: Zone.internalTestnet,
   transport,
 })
 const decoratedZoneClient = zoneClient.extend(decorator())
@@ -153,7 +153,7 @@ test('requestWithdrawalSync returns a receipt and sender tag', async () => {
   expectTypeOf(result.senderTag).toEqualTypeOf<Hash>()
 
   const explicitAccountClient = createClient({
-    chain: zoneModerato(7),
+    chain: Zone.internalTestnet,
     transport,
   })
   const explicitResult = await zoneActions.requestWithdrawalSync(
