+++ src/tempo/actions/zone.test.ts
@@ -22,7 +22,7 @@ import {
   writeContract,
 } from 'viem/actions'
 import { tempoModerato } from 'viem/chains'
-import { Abis, Actions } from 'viem/tempo'
+import { Abis, Actions, Addresses } from 'viem/tempo'
 import { parseUnits } from 'viem/utils'
 import { describe, expect, test } from 'vitest'
 import { accounts } from '~test/constants.js'
@@ -32,7 +32,6 @@ import { defineZone, zoneAdminKey } from '~test/tempo/prool.js'
 import {
   factoryAddress,
   getClient as getZoneClient,
-  portalAddress,
   unredactedRpcUrl,
   http as zoneHttp,
   zoneId,
@@ -40,14 +39,13 @@ import {
 import { createHttpServer } from '~test/utils.js'
 import * as WithdrawalSenderTag from '../internal/WithdrawalSenderTag.js'
 import * as Storage from '../Storage.js'
-import * as ZoneAbis from '../zones/Abis.js'
-import { getPortalAddress } from '../zones/zone.js'
 import * as tokenActions from './token.js'
 import * as zoneActions from './zone.js'
 
 const account = privateKeyToAccount(accounts[0].privateKey)
 const portalAdmin = privateKeyToAccount(zoneAdminKey)
 const tempoRefundRecipient = accounts[2].address
+const portalAddress = Addresses.zonePortal(zoneId)
 const mainnetClient = createClient({
   account,
   chain,
@@ -143,7 +141,7 @@ async function createUnconfiguredZone() {
       ? await writeContract(portalAdminClient, {
           account: portalAdmin,
           address: factoryAddress,
-          abi: ZoneAbis.zoneFactory,
+          abi: Abis.zoneFactory,
           functionName: 'createZone',
           args: [
             {
@@ -163,7 +161,7 @@ async function createUnconfiguredZone() {
       : await (async () => {
           const verifier = await readContract(mainnetClient, {
             address: factoryAddress,
-            abi: ZoneAbis.zoneFactory,
+            abi: Abis.zoneFactory,
             functionName: 'verifier',
           })
           const genesisTempoBlockNumber = BigInt(
@@ -172,7 +170,7 @@ async function createUnconfiguredZone() {
           return writeContract(mainnetClient, {
             account,
             address: factoryAddress,
-            abi: ZoneAbis.zoneFactory,
+            abi: Abis.zoneFactory,
             functionName: 'createZone',
             args: [
               {
@@ -193,7 +191,7 @@ async function createUnconfiguredZone() {
         })()
   const receipt = await waitForTransactionReceipt(mainnetClient, { hash })
   const [event] = parseEventLogs({
-    abi: ZoneAbis.zoneFactory,
+    abi: Abis.zoneFactory,
     eventName: 'ZoneCreated',
     logs: receipt.logs,
     strict: true,
@@ -210,7 +208,7 @@ async function getPortalCall(hash: Hash) {
   const transaction = await getTransaction(mainnetClient, { hash })
   const call = transaction.calls?.[1]
   if (!call?.data) throw new Error('Portal call is unavailable.')
-  return decodeFunctionData({ abi: ZoneAbis.zonePortal, data: call.data })
+  return decodeFunctionData({ abi: Abis.zonePortal, data: call.data })
 }
 
 describe('zone instance', () => {
@@ -230,7 +228,7 @@ describe('zone instance', () => {
         expect(sameZone).toBe(zone_)
         expect(zone_.zoneId).not.toBe(zoneId)
         expect(zone_.chainId).not.toBe(zoneClient.chain.id)
-        expect(zone_.portalAddress).not.toBe(portalAddress)
+        expect(Addresses.zonePortal(zone_.zoneId)).not.toBe(portalAddress)
 
         const response = await fetch(zone_.rpcUrl, {
           body: JSON.stringify({
@@ -580,7 +578,7 @@ describe('encryptedDeposit', () => {
       recipient: account.address,
       zoneId: 7,
     })
-    expect(registryCalls[1].address).toBe(getPortalAddress(tempoModerato.id, 7))
+    expect(registryCalls[1].address).toBe(Addresses.zonePortal(7))
 
     await expect(
       zoneActions.encryptedDeposit(mainnetClient, {
@@ -708,7 +706,6 @@ describe('deposit', () => {
   test('behavior: encodes Tempo refund recipient', () => {
     const parameters = {
       amount: 1n,
-      chainId: chain.id,
       portalAddress,
       recipient: account.address,
       tempoRefundRecipient: account.address,
@@ -726,10 +723,9 @@ describe('deposit', () => {
     const { portalAddress: _, ...registryParameters } = parameters
     const registryCalls = zoneActions.deposit.calls({
       ...registryParameters,
-      chainId: tempoModerato.id,
       zoneId: 7,
     })
-    expect(registryCalls[1].address).toBe(getPortalAddress(tempoModerato.id, 7))
+    expect(registryCalls[1].address).toBe(Addresses.zonePortal(7))
   })
 
   test.runIf(legacyZoneCallback)(
@@ -901,7 +897,7 @@ describe('requestWithdrawal', () => {
     if (!withdrawalCall?.data)
       throw new Error('Prepared withdrawal call is unavailable.')
     const decoded = decodeFunctionData({
-      abi: ZoneAbis.zoneOutbox,
+      abi: Abis.zoneOutbox,
       data: withdrawalCall.data,
     })
     expect(decoded.functionName).toBe('requestWithdrawal')
@@ -960,7 +956,7 @@ describe('requestWithdrawal', () => {
     expect(clientAccountResult.receipt).toEqual(clientAccountReceipt)
     expect(clientAccountResult.receipt.status).toBe('success')
     const [clientAccountEvent] = parseEventLogs({
-      abi: ZoneAbis.zoneOutbox,
+      abi: Abis.zoneOutbox,
       logs: clientAccountResult.receipt.logs,
       eventName: 'WithdrawalRequested',
       strict: true,
@@ -1000,7 +996,7 @@ describe('requestWithdrawal', () => {
     expect(explicitAccountResult.receipt).toEqual(explicitAccountReceipt)
     expect(explicitAccountResult.receipt.status).toBe('success')
     const [explicitAccountEvent] = parseEventLogs({
-      abi: ZoneAbis.zoneOutbox,
+      abi: Abis.zoneOutbox,
       logs: explicitAccountResult.receipt.logs,
       eventName: 'WithdrawalRequested',
       strict: true,
