+++ src/tempo/index.ts
@@ -35,11 +35,10 @@ export {
   type FallbackTransportErrorType,
   fallback,
 } from '../clients/transports/fallback.js'
-export {
-  type HttpTransport,
-  type HttpTransportConfig,
-  type HttpTransportErrorType,
-  http,
+export type {
+  HttpTransport,
+  HttpTransportConfig,
+  HttpTransportErrorType,
 } from '../clients/transports/http.js'
 export {
   type WebSocketTransport,
@@ -106,6 +105,13 @@ export type {
 } from './Transaction.js'
 export * as Transaction from './Transaction.js'
 export * as Transport from './Transport.js'
-export { walletNamespaceCompat, withFeePayer, withRelay } from './Transport.js'
+export {
+  type HttpConfig,
+  http,
+  walletNamespaceCompat,
+  withFeePayer,
+  withRelay,
+} from './Transport.js'
 export * as WebAuthnP256 from './WebAuthnP256.js'
 export * as WebCryptoP256 from './WebCryptoP256.js'
+export * as Zone from './Zone.js'
