--- src/tempo/zones/Abis.ts
@@ -1,1499 +1,0 @@
-// Generated with `pnpm gen:tempo-abis`. Do not modify manually.
-// Source: tempoxyz/zones@7e38642cdd5a1500784fd15cf30eda5366f3350a
-
-export const zoneFactory = [
-  {
-    name: 'owner',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'address' }],
-  },
-  {
-    name: 'transferOwnership',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'address', name: 'newOwner' }],
-    outputs: [],
-  },
-  {
-    name: 'createZone',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      {
-        type: 'tuple',
-        name: 'params',
-        components: [
-          { type: 'address', name: 'initialToken' },
-          { type: 'bool', name: 'accessMode' },
-          { type: 'bool', name: 'gatewayMode' },
-          { type: 'address[]', name: 'allowedAccounts' },
-          { type: 'address[]', name: 'zoneGateways' },
-          { type: 'address', name: 'admin' },
-          { type: 'address[]', name: 'sequencers' },
-          { type: 'uint8', name: 'threshold' },
-          { type: 'string', name: 'rpcUrl' },
-        ],
-      },
-    ],
-    outputs: [
-      { type: 'uint32', name: 'zoneId' },
-      { type: 'address', name: 'portal' },
-    ],
-  },
-  {
-    name: 'zones',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint32', name: 'zoneId' }],
-    outputs: [
-      {
-        type: 'tuple',
-        components: [
-          { type: 'uint32', name: 'zoneId' },
-          { type: 'address', name: 'portal' },
-          { type: 'bool', name: 'accessMode' },
-          { type: 'bool', name: 'gatewayMode' },
-          { type: 'address', name: 'admin' },
-          { type: 'address[]', name: 'sequencers' },
-          { type: 'uint8', name: 'threshold' },
-          { type: 'address', name: 'verifier' },
-          { type: 'string', name: 'rpcUrl' },
-        ],
-      },
-    ],
-  },
-  {
-    name: 'nextZoneId',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint32' }],
-  },
-  {
-    name: 'isZonePortal',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'address', name: 'portal' }],
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    name: 'OwnershipTransferred',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'previousOwner', indexed: true },
-      { type: 'address', name: 'newOwner', indexed: true },
-    ],
-  },
-  {
-    name: 'ZoneCreated',
-    type: 'event',
-    inputs: [
-      { type: 'uint32', name: 'zoneId', indexed: true },
-      { type: 'address', name: 'portal', indexed: true },
-      { type: 'address', name: 'initialToken' },
-      { type: 'bool', name: 'accessMode' },
-      { type: 'bool', name: 'gatewayMode' },
-      { type: 'address', name: 'admin' },
-      { type: 'address[]', name: 'sequencers' },
-      { type: 'uint8', name: 'threshold' },
-      { type: 'address', name: 'verifier' },
-    ],
-  },
-  { name: 'InvalidToken', type: 'error', inputs: [] },
-  { name: 'NotOwner', type: 'error', inputs: [] },
-  { name: 'InvalidAdmin', type: 'error', inputs: [] },
-  { name: 'InvalidSequencerSet', type: 'error', inputs: [] },
-  { name: 'InvalidClosedLoopConfig', type: 'error', inputs: [] },
-  { name: 'DuplicateAllowedAccount', type: 'error', inputs: [] },
-  { name: 'DuplicateZoneGateway', type: 'error', inputs: [] },
-  {
-    name: 'ZoneCreated',
-    type: 'event',
-    inputs: [
-      { type: 'uint32', name: 'zoneId', indexed: true },
-      { type: 'address', name: 'portal', indexed: true },
-      { type: 'address', name: 'messenger', indexed: true },
-      { type: 'address', name: 'initialToken' },
-      { type: 'address', name: 'admin' },
-      { type: 'address', name: 'sequencer' },
-      { type: 'address', name: 'verifier' },
-      { type: 'bytes32', name: 'genesisBlockHash' },
-      { type: 'bytes32', name: 'genesisTempoBlockHash' },
-      { type: 'uint64', name: 'genesisTempoBlockNumber' },
-    ],
-  },
-  {
-    name: 'createZone',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      {
-        type: 'tuple',
-        name: 'params',
-        components: [
-          { type: 'address', name: 'initialToken' },
-          { type: 'address', name: 'admin' },
-          { type: 'address', name: 'sequencer' },
-          { type: 'address', name: 'verifier' },
-          {
-            type: 'tuple',
-            name: 'zoneParams',
-            components: [
-              { type: 'bytes32', name: 'genesisBlockHash' },
-              { type: 'bytes32', name: 'genesisTempoBlockHash' },
-              { type: 'uint64', name: 'genesisTempoBlockNumber' },
-            ],
-          },
-          { type: 'string', name: 'rpcUrl' },
-        ],
-      },
-    ],
-    outputs: [
-      { type: 'uint32', name: 'zoneId' },
-      { type: 'address', name: 'portal' },
-    ],
-  },
-  {
-    name: 'verifier',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'address' }],
-  },
-] as const
-
-export const zonePortal = [
-  {
-    name: 'zoneId',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint32' }],
-  },
-  {
-    name: 'admin',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'address' }],
-  },
-  {
-    name: 'messenger',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'address' }],
-  },
-  {
-    name: 'isAccessEnforced',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    name: 'setAccessMode',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'bool', name: 'enforced' }],
-    outputs: [],
-  },
-  {
-    name: 'isGatewayOpen',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    name: 'setGatewayMode',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'bool', name: 'enforced' }],
-    outputs: [],
-  },
-  {
-    name: 'hasRole',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [
-      { type: 'address', name: 'account' },
-      { type: 'uint8', name: 'role' },
-    ],
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    name: 'setAllowedAccount',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'account' },
-      { type: 'bool', name: 'allowed' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'setGateway',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'account' },
-      { type: 'bool', name: 'allowed' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'setPauseGuardian',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'account' },
-      { type: 'bool', name: 'allowed' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'setSequencerSet',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address[]', name: 'newSequencers' },
-      { type: 'uint8', name: 'newThreshold' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'verifier',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'address' }],
-  },
-  {
-    name: 'sequencerSetVersion',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'sequencerThreshold',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint8' }],
-  },
-  {
-    name: 'zoneHeight',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'isSequencer',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'address', name: 'account' }],
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    name: 'sequencerCount',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'sequencerAt',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint256', name: 'index' }],
-    outputs: [{ type: 'address' }],
-  },
-  {
-    name: 'leader',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'address' }],
-  },
-  {
-    name: 'leaderEpoch',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'leaderActivationTempoBlock',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'setLeader',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'newLeader' },
-      { type: 'uint64', name: 'expectedEpoch' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'withdrawalBatchIndex',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'blockHash',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'bytes32' }],
-  },
-  {
-    name: 'currentDepositQueueHash',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'bytes32' }],
-  },
-  {
-    name: 'lastSyncedTempoBlockNumber',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'withdrawalQueueHead',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'withdrawalQueueTail',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'withdrawalQueueSlot',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint256', name: 'queueIndex' }],
-    outputs: [{ type: 'bytes32' }],
-  },
-  {
-    name: 'calculateDepositFee',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint128', name: 'fee' }],
-  },
-  {
-    name: 'calculateBouncebackFee',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint128', name: 'fee' }],
-  },
-  {
-    name: 'depositCount',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'lastProcessedDepositNumber',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'FIXED_DEPOSIT_GAS',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'MAX_GAS_FEE_RATE',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint128' }],
-  },
-  {
-    name: 'MAX_TOKENS_ENABLED_PER_TEMPO_BLOCK',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'MAX_TOKEN_METADATA_BYTES',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'areDepositsActive',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'address', name: 'token' }],
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    name: 'tokenConfig',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'address', name: 'token' }],
-    outputs: [
-      {
-        type: 'tuple',
-        components: [
-          { type: 'bool', name: 'enabled' },
-          { type: 'bool', name: 'depositsActive' },
-        ],
-      },
-    ],
-  },
-  {
-    name: 'initialize',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'uint32', name: 'zoneId' },
-      { type: 'address', name: 'initialToken' },
-      { type: 'bool', name: 'accessMode' },
-      { type: 'bool', name: 'gatewayMode' },
-      { type: 'address[]', name: 'allowedAccounts' },
-      { type: 'address[]', name: 'zoneGateways' },
-      { type: 'address', name: 'admin' },
-      { type: 'address', name: 'messenger' },
-      { type: 'address[]', name: 'sequencers' },
-      { type: 'uint8', name: 'threshold' },
-      { type: 'address', name: 'verifier' },
-      { type: 'string', name: 'rpcUrl' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'deliverWithdrawal',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'to' },
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'bytes32', name: 'memo' },
-      { type: 'uint64', name: 'gasLimit' },
-      { type: 'bytes', name: 'callbackData' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'MAX_DEPOSITS_PER_TEMPO_BLOCK',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'MAX_WITHDRAWAL_GAS_LIMIT',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'paused',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    name: 'pauseExpiry',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'abdicationEffectiveAt',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint8', name: 'capability' }],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'processWithdrawals',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      {
-        type: 'tuple[]',
-        name: 'withdrawals',
-        components: [
-          { type: 'address', name: 'token' },
-          { type: 'bytes32', name: 'senderTag' },
-          { type: 'address', name: 'to' },
-          { type: 'uint128', name: 'amount' },
-          { type: 'bytes32', name: 'memo' },
-          { type: 'uint64', name: 'gasLimit' },
-          { type: 'uint64', name: 'fallbackNonce' },
-          { type: 'bytes', name: 'callbackData' },
-          { type: 'bytes', name: 'encryptedSender' },
-        ],
-      },
-      { type: 'bytes32', name: 'remainingQueue' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'pause',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [],
-    outputs: [],
-  },
-  {
-    name: 'resume',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [],
-    outputs: [],
-  },
-  {
-    name: 'abdicate',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'uint8', name: 'capability' }],
-    outputs: [],
-  },
-  {
-    name: 'submitBatch',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'uint64', name: 'tempoBlockNumber' },
-      { type: 'uint64', name: 'recentTempoBlockNumber' },
-      {
-        type: 'tuple',
-        name: 'blockTransition',
-        components: [
-          { type: 'bytes32', name: 'prevBlockHash' },
-          { type: 'bytes32', name: 'nextBlockHash' },
-        ],
-      },
-      {
-        type: 'tuple',
-        name: 'depositQueueTransition',
-        components: [
-          { type: 'bytes32', name: 'prevProcessedHash' },
-          { type: 'bytes32', name: 'nextProcessedHash' },
-          { type: 'uint64', name: 'prevDepositNumber' },
-          { type: 'uint64', name: 'nextDepositNumber' },
-        ],
-      },
-      { type: 'bytes32', name: 'withdrawalQueueHash' },
-      { type: 'bytes', name: 'verifierConfig' },
-      { type: 'bytes', name: 'proof' },
-      { type: 'uint256', name: 'nextZoneHeight' },
-      { type: 'bytes[]', name: 'signatures' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'enableToken',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'address', name: 'token' }],
-    outputs: [],
-  },
-  {
-    name: 'pauseDeposits',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'address', name: 'token' }],
-    outputs: [],
-  },
-  {
-    name: 'resumeDeposits',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'address', name: 'token' }],
-    outputs: [],
-  },
-  {
-    name: 'setZoneGasRate',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'uint128', name: 'newZoneGasRate' }],
-    outputs: [],
-  },
-  {
-    name: 'setMaxTempoGasRate',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'uint128', name: 'newMaxTempoGasRate' }],
-    outputs: [],
-  },
-  {
-    name: 'setBouncebackGas',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'uint64', name: 'newBouncebackGas' }],
-    outputs: [],
-  },
-  {
-    name: 'transferAdmin',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'address', name: 'newAdmin' }],
-    outputs: [],
-  },
-  {
-    name: 'acceptAdmin',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [],
-    outputs: [],
-  },
-  {
-    name: 'rpcUrl',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'string' }],
-  },
-  {
-    name: 'setRpcUrl',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'string', name: 'rpcUrl' }],
-    outputs: [],
-  },
-  {
-    name: 'deposit',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'uint256', name: 'keyIndex' },
-      {
-        type: 'tuple',
-        name: 'encrypted',
-        components: [
-          { type: 'bytes32', name: 'ephemeralPubkeyX' },
-          { type: 'uint8', name: 'ephemeralPubkeyYParity' },
-          { type: 'bytes', name: 'ciphertext' },
-          { type: 'bytes12', name: 'nonce' },
-          { type: 'bytes16', name: 'tag' },
-        ],
-      },
-      { type: 'address', name: 'tempoRefundRecipient' },
-    ],
-    outputs: [{ type: 'bytes32', name: 'newCurrentDepositQueueHash' }],
-  },
-  {
-    name: 'depositEncrypted',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'uint256', name: 'keyIndex' },
-      {
-        type: 'tuple',
-        name: 'encrypted',
-        components: [
-          { type: 'bytes32', name: 'ephemeralPubkeyX' },
-          { type: 'uint8', name: 'ephemeralPubkeyYParity' },
-          { type: 'bytes', name: 'ciphertext' },
-          { type: 'bytes12', name: 'nonce' },
-          { type: 'bytes16', name: 'tag' },
-        ],
-      },
-      { type: 'address', name: 'tempoRefundRecipient' },
-    ],
-    outputs: [{ type: 'bytes32', name: 'newCurrentDepositQueueHash' }],
-  },
-  {
-    name: 'setSequencerEncryptionKey',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'bytes32', name: 'x' },
-      { type: 'uint8', name: 'yParity' },
-      { type: 'uint8', name: 'popV' },
-      { type: 'bytes32', name: 'popR' },
-      { type: 'bytes32', name: 'popS' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'isTokenEnabled',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'address', name: 'token' }],
-    outputs: [{ type: 'bool' }],
-  },
-  {
-    name: 'enabledTokenCount',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'enabledTokenAt',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint256', name: 'index' }],
-    outputs: [{ type: 'address' }],
-  },
-  {
-    name: 'tokenEnablementHash',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'bytes32' }],
-  },
-  {
-    name: 'zoneGasRate',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint128' }],
-  },
-  {
-    name: 'maxTempoGasRate',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint128' }],
-  },
-  {
-    name: 'bouncebackGas',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'pendingAdmin',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'address' }],
-  },
-  {
-    name: 'refunds',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [
-      { type: 'address', name: 'token' },
-      { type: 'address', name: 'owner' },
-    ],
-    outputs: [{ type: 'uint128' }],
-  },
-  {
-    name: 'sequencerEncryptionKey',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [
-      { type: 'bytes32', name: 'x' },
-      { type: 'uint8', name: 'yParity' },
-    ],
-  },
-  {
-    name: 'encryptionKeyCount',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'encryptionKeyAt',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint256', name: 'index' }],
-    outputs: [
-      {
-        type: 'tuple',
-        name: 'entry',
-        components: [
-          { type: 'bytes32', name: 'x' },
-          { type: 'uint8', name: 'yParity' },
-          { type: 'uint64', name: 'activationBlock' },
-        ],
-      },
-    ],
-  },
-  {
-    name: 'isEncryptionKeyValid',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint256', name: 'keyIndex' }],
-    outputs: [
-      { type: 'bool', name: 'valid' },
-      { type: 'uint64', name: 'expiresAtBlock' },
-    ],
-  },
-  {
-    name: 'encryptionKeyAtBlock',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint64', name: 'tempoBlockNumber' }],
-    outputs: [
-      { type: 'bytes32', name: 'x' },
-      { type: 'uint8', name: 'yParity' },
-      { type: 'uint256', name: 'keyIndex' },
-    ],
-  },
-  {
-    name: 'claimRefund',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'address', name: 'token' }],
-    outputs: [{ type: 'uint128', name: 'amount' }],
-  },
-  {
-    name: 'DepositMade',
-    type: 'event',
-    inputs: [
-      { type: 'bytes32', name: 'newCurrentDepositQueueHash', indexed: true },
-      { type: 'address', name: 'sender', indexed: true },
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'netAmount' },
-      { type: 'uint128', name: 'fee' },
-      { type: 'uint256', name: 'keyIndex' },
-      { type: 'bytes32', name: 'ephemeralPubkeyX' },
-      { type: 'uint8', name: 'ephemeralPubkeyYParity' },
-      { type: 'bytes', name: 'ciphertext' },
-      { type: 'bytes12', name: 'nonce' },
-      { type: 'bytes16', name: 'tag' },
-      { type: 'address', name: 'tempoRefundRecipient' },
-      { type: 'uint64', name: 'depositNumber' },
-    ],
-  },
-  {
-    name: 'TokenEnabled',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'token', indexed: true },
-      { type: 'string', name: 'name' },
-      { type: 'string', name: 'symbol' },
-      { type: 'string', name: 'currency' },
-    ],
-  },
-  {
-    name: 'DepositsPaused',
-    type: 'event',
-    inputs: [{ type: 'address', name: 'token', indexed: true }],
-  },
-  {
-    name: 'DepositsResumed',
-    type: 'event',
-    inputs: [{ type: 'address', name: 'token', indexed: true }],
-  },
-  {
-    name: 'PortalPaused',
-    type: 'event',
-    inputs: [{ type: 'address', name: 'account', indexed: true }],
-  },
-  {
-    name: 'PortalResumed',
-    type: 'event',
-    inputs: [{ type: 'address', name: 'account', indexed: true }],
-  },
-  {
-    name: 'AbdicationScheduled',
-    type: 'event',
-    inputs: [
-      { type: 'uint8', name: 'capability', indexed: true },
-      { type: 'uint64', name: 'effectiveAt' },
-    ],
-  },
-  {
-    name: 'RpcUrlUpdated',
-    type: 'event',
-    inputs: [{ type: 'string', name: 'rpcUrl' }],
-  },
-  {
-    name: 'SequencerEncryptionKeyUpdated',
-    type: 'event',
-    inputs: [
-      { type: 'bytes32', name: 'x' },
-      { type: 'uint8', name: 'yParity' },
-      { type: 'address', name: 'pubkey' },
-      { type: 'uint256', name: 'keyIndex' },
-      { type: 'uint64', name: 'activationBlock' },
-    ],
-  },
-  {
-    name: 'BatchSubmitted',
-    type: 'event',
-    inputs: [
-      { type: 'uint64', name: 'withdrawalBatchIndex', indexed: true },
-      { type: 'uint256', name: 'withdrawalQueueIndex', indexed: true },
-      { type: 'bytes32', name: 'nextProcessedDepositQueueHash' },
-      { type: 'bytes32', name: 'nextBlockHash' },
-      { type: 'bytes32', name: 'withdrawalQueueHash' },
-      { type: 'uint64', name: 'lastProcessedDepositNumber' },
-    ],
-  },
-  {
-    name: 'WithdrawalProcessed',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'to', indexed: true },
-      { type: 'bytes32', name: 'senderTag', indexed: true },
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'bool', name: 'callbackSuccess' },
-    ],
-  },
-  {
-    name: 'WithdrawalBounceBack',
-    type: 'event',
-    inputs: [
-      { type: 'bytes32', name: 'newCurrentDepositQueueHash', indexed: true },
-      { type: 'uint64', name: 'fallbackNonce', indexed: true },
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'uint64', name: 'depositNumber' },
-    ],
-  },
-  {
-    name: 'DepositBounceBack',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'tempoRefundRecipient', indexed: true },
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'uint128', name: 'bouncebackFee' },
-    ],
-  },
-  {
-    name: 'DepositBounceBackPending',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'tempoRefundRecipient', indexed: true },
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'uint128', name: 'bouncebackFee' },
-    ],
-  },
-  {
-    name: 'RefundClaimed',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'recipient', indexed: true },
-      { type: 'address', name: 'token', indexed: true },
-      { type: 'uint128', name: 'amount' },
-    ],
-  },
-  {
-    name: 'ZoneGasRateUpdated',
-    type: 'event',
-    inputs: [{ type: 'uint128', name: 'zoneGasRate' }],
-  },
-  {
-    name: 'MaxTempoGasRateUpdated',
-    type: 'event',
-    inputs: [{ type: 'uint128', name: 'maxTempoGasRate' }],
-  },
-  {
-    name: 'BouncebackGasUpdated',
-    type: 'event',
-    inputs: [{ type: 'uint64', name: 'bouncebackGas' }],
-  },
-  {
-    name: 'AdminTransferStarted',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'currentAdmin', indexed: true },
-      { type: 'address', name: 'pendingAdmin', indexed: true },
-    ],
-  },
-  {
-    name: 'AdminTransferred',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'previousAdmin', indexed: true },
-      { type: 'address', name: 'newAdmin', indexed: true },
-    ],
-  },
-  {
-    name: 'RoleUpdated',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'account', indexed: true },
-      { type: 'uint8', name: 'prev' },
-      { type: 'uint8', name: 'next' },
-    ],
-  },
-  {
-    name: 'EnforcementModesUpdated',
-    type: 'event',
-    inputs: [
-      { type: 'bool', name: 'accessMode' },
-      { type: 'bool', name: 'gatewayMode' },
-    ],
-  },
-  {
-    name: 'SequencerSetUpdated',
-    type: 'event',
-    inputs: [
-      { type: 'uint64', name: 'nonce', indexed: true },
-      { type: 'uint8', name: 'threshold' },
-      { type: 'address[]', name: 'sequencers' },
-    ],
-  },
-  {
-    name: 'LeaderUpdated',
-    type: 'event',
-    inputs: [
-      { type: 'address', name: 'previousLeader', indexed: true },
-      { type: 'address', name: 'newLeader', indexed: true },
-      { type: 'uint64', name: 'epoch', indexed: true },
-      { type: 'uint64', name: 'activationTempoBlock' },
-    ],
-  },
-  { name: 'NotSequencer', type: 'error', inputs: [] },
-  { name: 'NotAdmin', type: 'error', inputs: [] },
-  { name: 'NotPauseAuthority', type: 'error', inputs: [] },
-  {
-    name: 'CapabilityAbdicated',
-    type: 'error',
-    inputs: [{ type: 'uint8', name: 'capability' }],
-  },
-  {
-    name: 'AbdicationAlreadyScheduled',
-    type: 'error',
-    inputs: [{ type: 'uint8', name: 'capability' }],
-  },
-  { name: 'PortalIsPaused', type: 'error', inputs: [] },
-  { name: 'NotPendingAdmin', type: 'error', inputs: [] },
-  { name: 'InvalidProof', type: 'error', inputs: [] },
-  { name: 'InvalidTempoBlockNumber', type: 'error', inputs: [] },
-  { name: 'NotFactory', type: 'error', inputs: [] },
-  { name: 'NotSelf', type: 'error', inputs: [] },
-  { name: 'AlreadyInitialized', type: 'error', inputs: [] },
-  { name: 'MustDelegateCall', type: 'error', inputs: [] },
-  { name: 'CallbackRejected', type: 'error', inputs: [] },
-  { name: 'TransferFailed', type: 'error', inputs: [] },
-  { name: 'ReentrantWithdrawal', type: 'error', inputs: [] },
-  {
-    name: 'EncryptionKeyExpired',
-    type: 'error',
-    inputs: [
-      { type: 'uint256', name: 'keyIndex' },
-      { type: 'uint64', name: 'activationBlock' },
-      { type: 'uint64', name: 'supersededAtBlock' },
-    ],
-  },
-  {
-    name: 'InvalidEncryptionKeyIndex',
-    type: 'error',
-    inputs: [{ type: 'uint256', name: 'keyIndex' }],
-  },
-  { name: 'NoEncryptionKeySet', type: 'error', inputs: [] },
-  {
-    name: 'NoEncryptionKeyAtBlock',
-    type: 'error',
-    inputs: [{ type: 'uint64', name: 'blockNumber' }],
-  },
-  { name: 'InvalidEphemeralPubkey', type: 'error', inputs: [] },
-  {
-    name: 'InvalidCiphertextLength',
-    type: 'error',
-    inputs: [
-      { type: 'uint256', name: 'actual' },
-      { type: 'uint256', name: 'expected' },
-    ],
-  },
-  { name: 'InvalidProofOfPossession', type: 'error', inputs: [] },
-  { name: 'DepositTooSmall', type: 'error', inputs: [] },
-  {
-    name: 'TokenEnablementBlockCapacityExceeded',
-    type: 'error',
-    inputs: [{ type: 'uint64', name: 'maximum' }],
-  },
-  { name: 'TokenMetadataTooLong', type: 'error', inputs: [] },
-  { name: 'GasFeeRateTooHigh', type: 'error', inputs: [] },
-  { name: 'DepositsNotActive', type: 'error', inputs: [] },
-  { name: 'TokenAlreadyEnabled', type: 'error', inputs: [] },
-  { name: 'TokenTransferPolicyNotSet', type: 'error', inputs: [] },
-  { name: 'InvalidDepositTransition', type: 'error', inputs: [] },
-  { name: 'InvalidSequencerSet', type: 'error', inputs: [] },
-  { name: 'SequencerConfigurationUnchanged', type: 'error', inputs: [] },
-  { name: 'InvalidQuorumCertificate', type: 'error', inputs: [] },
-  { name: 'CallbackDidNotReturnToZone', type: 'error', inputs: [] },
-  { name: 'InvalidBouncebackRecipient', type: 'error', inputs: [] },
-  { name: 'TokenNotEnabled', type: 'error', inputs: [] },
-  {
-    name: 'DepositBlockCapacityExceeded',
-    type: 'error',
-    inputs: [{ type: 'uint64', name: 'maximum' }],
-  },
-  { name: 'InvalidCallbackTarget', type: 'error', inputs: [] },
-  {
-    name: 'AccountNotAllowed',
-    type: 'error',
-    inputs: [{ type: 'address', name: 'account' }],
-  },
-  { name: 'InvalidLeader', type: 'error', inputs: [] },
-  { name: 'ActiveLeaderRemoved', type: 'error', inputs: [] },
-  { name: 'LeaderAlreadyUpdatedThisBlock', type: 'error', inputs: [] },
-  {
-    name: 'StaleLeadershipEpoch',
-    type: 'error',
-    inputs: [
-      { type: 'uint64', name: 'expected' },
-      { type: 'uint64', name: 'actual' },
-    ],
-  },
-  {
-    name: 'deposit',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: '_token' },
-      { type: 'address', name: 'to' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'bytes32', name: 'memo' },
-      { type: 'address', name: 'tempoRefundRecipient' },
-    ],
-    outputs: [{ type: 'bytes32' }],
-  },
-] as const
-
-export const zoneOutbox = [
-  {
-    name: 'tempoGasRate',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint128' }],
-  },
-  {
-    name: 'maxWithdrawalsPerBlock',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint32' }],
-  },
-  {
-    name: 'lastBatch',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [
-      {
-        type: 'tuple',
-        components: [
-          { type: 'bytes32', name: 'withdrawalQueueHash' },
-          { type: 'uint64', name: 'withdrawalBatchIndex' },
-        ],
-      },
-    ],
-  },
-  {
-    name: 'lastFinalizedTimestamp',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'nextWithdrawalIndex',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'lastFallbackNonce',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'pendingWithdrawalsCount',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'getPendingWithdrawals',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [
-      {
-        type: 'tuple[]',
-        components: [
-          { type: 'address', name: 'token' },
-          { type: 'address', name: 'sender' },
-          { type: 'bytes32', name: 'txHash' },
-          { type: 'address', name: 'to' },
-          { type: 'uint128', name: 'amount' },
-          { type: 'bytes32', name: 'memo' },
-          { type: 'uint64', name: 'gasLimit' },
-          { type: 'uint64', name: 'fallbackNonce' },
-          { type: 'bytes', name: 'callbackData' },
-          { type: 'bytes', name: 'revealTo' },
-        ],
-      },
-    ],
-  },
-  {
-    name: 'consumeFallbackRecipient',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'uint64', name: 'fallbackNonce' }],
-    outputs: [{ type: 'address', name: 'recipient' }],
-  },
-  {
-    name: 'calculateWithdrawalFee',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [{ type: 'uint64', name: 'gasLimit' }],
-    outputs: [{ type: 'uint128', name: 'fee' }],
-  },
-  {
-    name: 'MAX_CALLBACK_DATA_SIZE',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'MAX_WITHDRAWAL_GAS_LIMIT',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'WITHDRAWAL_BASE_GAS',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint64' }],
-  },
-  {
-    name: 'REVEAL_TO_KEY_LENGTH',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'AUTHENTICATED_WITHDRAWAL_CIPHERTEXT_LENGTH',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [],
-    outputs: [{ type: 'uint256' }],
-  },
-  {
-    name: 'setTempoGasRate',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'uint128', name: '_tempoGasRate' }],
-    outputs: [],
-  },
-  {
-    name: 'setMaxWithdrawalsPerBlock',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [{ type: 'uint32', name: '_maxWithdrawalsPerBlock' }],
-    outputs: [],
-  },
-  {
-    name: 'requestWithdrawal',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'token' },
-      { type: 'address', name: 'to' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'bytes32', name: 'memo' },
-      { type: 'uint64', name: 'gasLimit' },
-      { type: 'address', name: 'zoneFallbackRecipient' },
-      { type: 'bytes', name: 'data' },
-      { type: 'bytes', name: 'revealTo' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'enqueueDepositBounceBack',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'address', name: 'token' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'address', name: 'tempoRefundRecipient' },
-    ],
-    outputs: [],
-  },
-  {
-    name: 'finalizeWithdrawalBatch',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'uint256', name: 'count' },
-      { type: 'uint64', name: 'blockNumber' },
-      { type: 'bytes[]', name: 'encryptedSenders' },
-    ],
-    outputs: [{ type: 'bytes32', name: 'withdrawalQueueHash' }],
-  },
-  {
-    name: 'WithdrawalRequested',
-    type: 'event',
-    inputs: [
-      { type: 'uint64', name: 'withdrawalIndex', indexed: true },
-      { type: 'address', name: 'sender', indexed: true },
-      { type: 'address', name: 'token' },
-      { type: 'address', name: 'to' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'uint128', name: 'fee' },
-      { type: 'bytes32', name: 'memo' },
-      { type: 'uint64', name: 'gasLimit' },
-      { type: 'uint64', name: 'fallbackNonce' },
-      { type: 'bytes', name: 'data' },
-      { type: 'bytes', name: 'revealTo' },
-    ],
-  },
-  {
-    name: 'BatchFinalized',
-    type: 'event',
-    inputs: [
-      { type: 'bytes32', name: 'withdrawalQueueHash', indexed: true },
-      { type: 'uint64', name: 'withdrawalBatchIndex' },
-    ],
-  },
-  {
-    name: 'TempoGasRateUpdated',
-    type: 'event',
-    inputs: [{ type: 'uint128', name: 'tempoGasRate' }],
-  },
-  {
-    name: 'MaxWithdrawalsPerBlockUpdated',
-    type: 'event',
-    inputs: [{ type: 'uint32', name: 'maxWithdrawalsPerBlock' }],
-  },
-  { name: 'OnlySequencer', type: 'error', inputs: [] },
-  { name: 'GasLimitTooHigh', type: 'error', inputs: [] },
-  { name: 'OnlyZoneInbox', type: 'error', inputs: [] },
-  {
-    name: 'InvalidWithdrawalCount',
-    type: 'error',
-    inputs: [
-      { type: 'uint256', name: 'actual' },
-      { type: 'uint256', name: 'expected' },
-    ],
-  },
-  {
-    name: 'InvalidEncryptedSenderCount',
-    type: 'error',
-    inputs: [
-      { type: 'uint256', name: 'actual' },
-      { type: 'uint256', name: 'expected' },
-    ],
-  },
-  {
-    name: 'InvalidEncryptedSenderLength',
-    type: 'error',
-    inputs: [
-      { type: 'uint256', name: 'actual' },
-      { type: 'uint256', name: 'expected' },
-    ],
-  },
-  { name: 'InvalidFallbackRecipient', type: 'error', inputs: [] },
-  { name: 'CallbackDataTooLarge', type: 'error', inputs: [] },
-  { name: 'GasFeeRateTooHigh', type: 'error', inputs: [] },
-  { name: 'TransferFailed', type: 'error', inputs: [] },
-  { name: 'InvalidBlockNumber', type: 'error', inputs: [] },
-  { name: 'TooManyWithdrawalsThisBlock', type: 'error', inputs: [] },
-  { name: 'InvalidRevealTo', type: 'error', inputs: [] },
-  { name: 'InvalidCurrentTxHash', type: 'error', inputs: [] },
-  { name: 'ZeroAmountWithdrawal', type: 'error', inputs: [] },
-  { name: 'StaticCallNotAllowed', type: 'error', inputs: [] },
-] as const
-
-export const zoneMessenger = [
-  {
-    name: 'relayMessage',
-    type: 'function',
-    stateMutability: 'nonpayable',
-    inputs: [
-      { type: 'uint32', name: 'zoneId' },
-      { type: 'address', name: 'token' },
-      { type: 'bytes32', name: 'senderTag' },
-      { type: 'address', name: 'target' },
-      { type: 'uint128', name: 'amount' },
-      { type: 'uint64', name: 'gasLimit' },
-      { type: 'bytes', name: 'data' },
-    ],
-    outputs: [],
-  },
-] as const
-
-export const zoneVerifier = [
-  {
-    name: 'verify',
-    type: 'function',
-    stateMutability: 'view',
-    inputs: [
-      { type: 'uint32', name: 'zoneId' },
-      { type: 'uint64', name: 'tempoBlockNumber' },
-      { type: 'uint64', name: 'anchorBlockNumber' },
-      { type: 'bytes32', name: 'anchorBlockHash' },
-      { type: 'uint64', name: 'expectedWithdrawalBatchIndex' },
-      {
-        type: 'tuple',
-        name: 'blockTransition',
-        components: [
-          { type: 'bytes32', name: 'prevBlockHash' },
-          { type: 'bytes32', name: 'nextBlockHash' },
-        ],
-      },
-      {
-        type: 'tuple',
-        name: 'depositQueueTransition',
-        components: [
-          { type: 'bytes32', name: 'prevProcessedHash' },
-          { type: 'bytes32', name: 'nextProcessedHash' },
-          { type: 'uint64', name: 'prevDepositNumber' },
-          { type: 'uint64', name: 'nextDepositNumber' },
-        ],
-      },
-      { type: 'bytes32', name: 'withdrawalQueueHash' },
-      { type: 'bytes', name: 'verifierConfig' },
-      { type: 'bytes', name: 'proof' },
-    ],
-    outputs: [{ type: 'bool' }],
-  },
-] as const
