--- src/tempo/zones/index.ts
@@ -1,10 +1,0 @@
-// biome-ignore lint/performance/noBarrelFile: entrypoint module
-export * as Abis from './Abis.js'
-export * as Addresses from './Addresses.js'
-export { http, type ZoneHttpConfig } from './transport.js'
-export {
-  from,
-  getPortalAddress,
-  zone,
-  zoneModerato,
-} from './zone.js'
